Improved typing and bindings usage
