+++ CMakeLists.txt
@@ -12,6 +12,15 @@ if (DEFINED PY_BUILD_CMAKE_PACKAGE_VERSION)
     endif ()
 endif ()
 
+# set variables
+
+if (DEFINED PY_BUILD_CMAKE_MODULE_NAME)
+    set(STRUCTSTORE_INSTALL_DIR ${PY_BUILD_CMAKE_MODULE_NAME})
+else ()
+    set(CMAKE_INSTALL_PREFIX ${PROJECT_SOURCE_DIR})
+    set(STRUCTSTORE_INSTALL_DIR install)
+endif ()
+
 # dependencies
 
 find_package(Python 3.9 COMPONENTS Interpreter Development REQUIRED)
@@ -30,6 +39,7 @@ file(GLOB STRUCTSTORE_INC_FILES
         ${PROJECT_SOURCE_DIR}/include/*.hpp)
 
 set(STRUCTSTORE_SRC_FILES
+        ${PROJECT_SOURCE_DIR}/src/stst_bindings.cpp
         ${PROJECT_SOURCE_DIR}/src/stst_containers.cpp
         ${PROJECT_SOURCE_DIR}/src/stst_field.cpp
         ${PROJECT_SOURCE_DIR}/src/stst_shared.cpp
@@ -42,14 +52,15 @@ set(STRUCTSTORE_INC_DIRS
 
 # targets: library and python bindings
 
-add_library(structstore_lib STATIC
+add_library(structstore_lib SHARED
         ${STRUCTSTORE_INC_FILES}
         ${STRUCTSTORE_SRC_FILES})
 set_target_properties(structstore_lib PROPERTIES
         ARCHIVE_OUTPUT_DIRECTORY ${PROJECT_SOURCE_DIR}/lib
         ARCHIVE_OUTPUT_NAME structstore)
+target_link_libraries(structstore_lib PUBLIC nanobind-static)
 
-nanobind_add_module(structstore_bindings src/stst_bindings.cpp)
+nanobind_add_module(structstore_bindings src/structstore_bindings.cpp NB_STATIC)
 target_link_libraries(structstore_bindings PRIVATE structstore_lib)
 set_target_properties(structstore_bindings PROPERTIES
         LIBRARY_OUTPUT_DIRECTORY ${PROJECT_SOURCE_DIR}/lib
@@ -65,27 +76,23 @@ set_target_properties(structstore_bindings PROPERTIES
 if (CMAKE_SYSTEM_NAME MATCHES "Linux")
     target_link_options(structstore_bindings PRIVATE "LINKER:--exclude-libs,ALL")
 endif ()
+set_target_properties(structstore_bindings PROPERTIES
+        INSTALL_RPATH "\\\$ORIGIN")
 
 foreach (TARGET structstore_lib structstore_bindings)
     target_compile_features(${TARGET} PUBLIC cxx_std_17)
     target_compile_options(${TARGET} PUBLIC -fPIC)
     target_compile_options(${TARGET} PUBLIC -Wall -Wextra -pedantic)
-    target_include_directories(${TARGET} PRIVATE ${STRUCTSTORE_INC_DIRS})
-    target_link_libraries(${TARGET} PRIVATE
+    target_include_directories(${TARGET}
+            PRIVATE ${STRUCTSTORE_INC_DIRS})
+    target_link_libraries(${TARGET} PUBLIC
             ${YAML_CPP_LIBRARIES}
             rt)
 endforeach ()
 
 # install
 
-if (DEFINED PY_BUILD_CMAKE_MODULE_NAME)
-    set(STRUCTSTORE_INSTALL_DIR ${PY_BUILD_CMAKE_MODULE_NAME})
-else ()
-    set(CMAKE_INSTALL_PREFIX ${PROJECT_SOURCE_DIR})
-    set(STRUCTSTORE_INSTALL_DIR install)
-endif ()
-
-install(TARGETS structstore_lib
+install(TARGETS structstore_lib nanobind-static
         EXPORT StructStoreTargets
         COMPONENT python_modules
         DESTINATION ${STRUCTSTORE_INSTALL_DIR})
@@ -101,7 +108,7 @@ nanobind_stubgen(structstore_bindings)
 nanobind_stubgen_install(structstore_bindings ${STRUCTSTORE_INSTALL_DIR})
 
 # inspired by https://stackoverflow.com/a/49857699
-export(TARGETS structstore_lib
+export(TARGETS structstore_lib nanobind-static
         FILE "${PROJECT_BINARY_DIR}/StructStoreTargets.cmake")
 export(PACKAGE StructStore)
 configure_file(${PROJECT_SOURCE_DIR}/cmake/StructStoreConfig.cmake.in
